Add upload dialog to systems and catalogue items #1026
